Corrección propiedades de clases.

+++ src/main/java/com/ccsw/bidoffice/filetype/model/FileTypeDto.java
@@ -6,7 +6,7 @@ public class FileTypeDto {
 
     private String name;
 
-    private Long priority;
+    private Integer priority;
 
     public Long getId() {
         return id;
@@ -24,11 +24,11 @@ public class FileTypeDto {
         this.name = name;
     }
 
-    public Long getPriority() {
+    public Integer getPriority() {
         return priority;
     }
 
-    public void setPriority(Long priority) {
+    public void setPriority(Integer priority) {
         this.priority = priority;
     }
